Problem with repeats in entertainer.krn
I think there's a problem with the labeling of sections.  There is a 4-bar phrase (bars 72-75, inclusive) that is encoded as being the beginning of section E (which is the last strain of the piece), which is not correct --- it should be encoded as the last part of section D2 instead.  I think this is clear from the sheet music---there is a start repeat section indicated at measure 76, not 72, so section E should start at 76, not 72.  This problem manifests when repeats are automatically unrolled (which the kernscores website has the ability to do), and what happens is measures 72-75 become part of the repeated section E, when they really shouldn't.

Anyway, I think this change fixes the problem by moving `*>E	*>E	*>E` from bar 72 to 76.